Develop (#3)
* Working on an overhaul of the entire system, reducing the focus on
user/user data exchange and strengthening the focus on user/app data
interactions.

The system should no longer try to check whether users are members of 
the same group, but instead focus on whether applications can access
the data they provide to PHPAS for those applications.

* Working to provide the whole App Permission ordeal.

* Working on the new safer app authentication mechanism. And on the
newly revisited UI.

* Working on the new features and layout for the next release.

* Working on getting the system to a releaseable status.

* The permissions can now be properly adjusted by the user. The only
thing needed is to also apply them when the application reads these.

* I've manily made progress on the domain blocking features. I am
considering revamping the way the system works for CIDR storage,
while it should work for the purpose it serves, it's far from perfect

* Touched up the user profiles. I am currently working on an
implementation to micro-serviced webhooks that does not
suck when trying to work with it.

The service needs to be able to access the list of apps installed
on the server to allow an administator using the new hook app to
connect to applications from the list.

It would also be recommended to have a way that allows applications
integrated with the server to provide a list of hooks that they
support. This way a user could potentially manage the hooks from
the UI. Probably, the smartest way to achieve this is by letting
the app upload a JSON file with supported hooks, captions and
descriptions (including hierarchies)

On the other hand, apps should be able to create hooks themselves
by registering them or checking whether the hooks they need already
do exist.

Applications should be able to request "transliteration".
Sometimes the payload from one app does not fit the API of the
receiving app, when the vendor of one of the apps is not ourselves,
this could could become a problem if the server does not allow
translating the formats slightly.

* PHPAS is slowly getting ready to be pushed. I'd like to brush up a little
issue I have with the contexts. Once that is done I can deploy it.

* Working on the webhook subsystem. I've been able to externalize the
behavior towards CptnH00k and need to address some minor behavior
inconsistencies.

* Renamed the "src" app to local

* Fixed a function name that PHP7.0 does not accept but 7.1 does.
Seems like too minor an issue to not fix.

Added a bit of documentation to the context.

* Webhook changes.

A lot of progress has been made towards removing the old webhook 
mechanism in favor of the new external one.

Tasks that are open before I can consider the webhooks to be done:

1. The UI must be available for the webhook management software. Right
   now CptnH00k is absolutely headless, styleless and unstable.

2. The integrations need to be more stable. The hooks for CptnH00k 
   should be disabled in the interface, in turn, the SSO hooks should
   be listed somewhere else.

3. The application detail page makes no sense. The secret and appId are
   shown separately and in seemingly random places, the UI implies that
   a record, once modified is stored, but it requires the user to click
   "store"

* Working on the hook integration.

+++ bin/controllers/app.php
@@ -1,9 +1,17 @@
 <?php
 
-use spitfire\exceptions\PrivateException;
 use spitfire\exceptions\PublicException;
 use spitfire\io\Upload;
+use spitfire\storage\database\pagination\Paginator;
 
+/**
+ * This controller allows administrators (and only those) to manage the applications
+ * that can connect to the server and manage their preferences and default
+ * access level settings.
+ * 
+ * Only admins receive access since this is the strongest vector for a malicious
+ * application to raise it's privileges and access data it's not supposed to have.
+ */
 class AppController extends BaseController
 {
 		
@@ -21,22 +29,10 @@ class AppController extends BaseController
 	public function index() {
 		
 		$query = db()->table('authapp')->getAll();
-		$pag   = new Pagination($query, 'app');
+		$pag   = new Paginator($query);
 		
-		$this->view->set('query',      $query);
 		$this->view->set('pagination', $pag);
 		
-	}
-
-	static private function _getRandomBytes($length, &$crypto_strong = null){
-		if (function_exists('random_bytes')){
-			$crypto_strong = true;
-			return random_bytes($length);
-		}
-		else {
-			$crypto_strong = false;
-			return openssl_random_pseudo_bytes($length, $crypto_strong);
-		}
 	}
 	
 	public function create() {
@@ -44,11 +40,9 @@ class AppController extends BaseController
 		if ($this->request->isPost()) {
 			$app = db()->table('authapp')->newRecord();
 			$app->name      = $_POST['name'];
-			$app->appSecret = preg_replace('/[^a-z\d]/i', '', base64_encode(self::_getRandomBytes(35, $secure)));
-			
-			if (!$secure) {
-				throw new PrivateException('Could not generate safe AppSecret');
-			}
+			$app->appSecret = preg_replace('/[^a-z\d]/i', '', base64_encode(random_bytes(35)));
+			$app->system    = false;
+			$app->drawer    = false;
 			
 			if ($_POST['icon'] instanceof Upload) {
 				$app->icon = $_POST['icon']->validate()->store();
@@ -60,15 +54,13 @@ class AppController extends BaseController
 			} while ($count !== 0);
 			
 			$app->store();
-			$this->response->getHeaders()->redirect(new URL('app', 'index', null, Array('message' => 'success')));
+			$this->response->getHeaders()->redirect(url('app', 'index', Array('message' => 'success')));
 			return;
 		}
 		
 	}
 	
-	public function detail($appID) {
-		
-		$app = db()->table('authapp')->get('_id', $appID)->fetch();
+	public function detail(AuthAppModel$app) {
 		
 		if ($this->request->isPost()) {
 			
@@ -83,41 +75,37 @@ class AppController extends BaseController
 			}
 			
 			if ($_POST['icon'] instanceof Upload) {
-				$app->icon = $_POST['icon']->validate()->store();
+				$app->icon = $_POST['icon']->store();
 			}
+			
+			$app->system = isset($_POST['system']);
+			$app->drawer = isset($_POST['drawer']);
 			
 			$app->store();
 		}
 		
 		$this->view->set('app', $app);
+		
+		try {
+			$hookapp = db()->table('authapp')->get('_id', SysSettingModel::getValue('cptn.h00k'))->first(true)->appID;
+			$this->view->set('webhooks', $this->hook->on($hookapp, $app->appID)->listeners);
+		} catch (Exception $ex) {
+			$this->view->set('webhooks', []);
+		}
 	}
 	
 	public function delete($appID) {
+		$xsrf = new \spitfire\io\XSSToken();
 		
-		if (isset($_GET['confirm'])) {
+		if (isset($_GET['confirm']) && $xsrf->verify($_GET['confirm'])) {
 			$app = db()->table('authapp')->get('_id', $appID)->fetch();
 			$app->delete();
 			
-			$this->response->getHeaders()->redirect(new URL('app', 'index', null, Array('message' => 'deleted')));
+			$this->response->getHeaders()->redirect(url('app', 'index', Array('message' => 'deleted')));
 			return;
 		}
 		
-		$this->view->set('confirm', url('app', 'delete', $appID, Array('confirm' => 'true')));
-	}
-	
-	/**
-	 * 
-	 * @template none
-	 * @layout none
-	 * @param type $appID
-	 */
-	public function deauthorize($appID) {
-		$app  = db()->table('authapp')->get('_id', $appID)->fetch();
-		$auth = db()->table('user\authorizedapp')->get('user', $this->user)->addRestriction('app', $app)->fetch();
-		
-		if ($auth) { $auth->delete(); }
-		
-		$this->response->getHeaders()->redirect(url());
+		$this->view->set('confirm', url('app', 'delete', $appID, Array('confirm' => $xsrf->getValue())));
 	}
 	
 }
